server/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python
scikit-learn
PyWavelets
joblib
python-dotenv